Switched to Byte instead of Short

+++ Midterm Project/src/Main.java
@@ -1,31 +1,31 @@
 import java.util.*;
 
 public class Main {
-    static final short MAX = 6;
-    static short GUESSES = 12;
+    static final Byte MAX = 6;
+    static Byte GUESSES = 12;
     static final String REGEX = "^[1-6] [1-6] [1-6] [1-6]$";
 
-    static List<Short> numberBank = new ArrayList<>();
+    static List<Byte> numberBank = new ArrayList<>();
     static {
-        for (short i = 1; i <= MAX; i++) {
+        for (Byte i = 1; i <= MAX; i++) {
             numberBank.add(i);
         }
     }
 
-    public static List<Short> newCode() {
+    public static List<Byte> newCode() {
         Collections.shuffle(numberBank);
         return numberBank.subList(0, 4);
     }
 
     @SuppressWarnings("unused")
-    public static void printCode(List<Short> code) {
-        for (Short number : code) {
+    public static void printCode(List<Byte> code) {
+        for (Byte number : code) {
             System.out.print(number + " ");
         }
         System.out.println();
     }
 
-    public static List<Short> newGuess(String prompt) {
+    public static List<Byte> newGuess(String prompt) {
         while (true) {
             System.out.print(prompt);
             String input = new Scanner(System.in).nextLine();
@@ -33,16 +33,16 @@ public class Main {
                 System.out.println("Invalid input. Please try again.");
             }
             else {
-                List<Short> guess = new ArrayList<>();
+                List<Byte> guess = new ArrayList<>();
                 for (String number : input.split(" ")) {
-                    guess.add(Short.parseShort(number.trim()));
+                    guess.add(Byte.parseByte(number.trim()));
                 }
                 return guess;
             }
         }
     }
 
-    public static String testGuess(List<Short> guess, List<Short> secret) {
+    public static String testGuess(List<Byte> guess, List<Byte> secret) {
         StringBuilder output = new StringBuilder();
 
         for (int i = 0; i < secret.size(); i++) {
@@ -60,10 +60,10 @@ public class Main {
     }
 
     public static void main(String[] args) {
-        List<Short> secret = newCode();
+        List<Byte> secret = newCode();
 
         while (GUESSES > 0) {
-            List<Short> guess = newGuess("Type your guess >> ");
+            List<Byte> guess = newGuess("Type your guess >> ");
             String output = testGuess(guess, secret);
             System.out.println();
